+++ src/components/ParticipantsList/index.tsx
@@ -1,5 +1,5 @@
 import ParticipantCard from "@components/ParticipantCard";
-import { useChatDataContext } from "@components/context/СhatDataContext";
+import { useChatDataContext } from "src/contexts/СhatDataContext";
 import type { IUser } from "@app-types/user";
 import { type ReactElement } from "react";
